execs

+++ lista 1 recursividade/exe003.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int soma(int vet[], int tam){
+int soma(float vet[], int tam){
     if(tam == -1){
         return 0;
     }else{
@@ -10,10 +10,10 @@ int main(){
     int tam;
     printf("Digite o tamanho do vetor: ");
     scanf("%i", &tam);
-    int vet[tam];
+    float vet[tam];
     for (int i = 0; i < tam; i++){
         printf("Digite o valor da posicao %i: ", i);
-        scanf("%i", &vet[i]);
+        scanf("%f", &vet[i]);
     }
     printf("A soma dos valores e %i", soma(vet, tam - 1));
 }
